+++ ft_memcpy.c
@@ -6,15 +6,15 @@
 /*   By: vtarasov <vtarasov@student.42warsaw.pl>    +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2026/06/16 14:36:55 by vtarasov          #+#    #+#             */
-/*   Updated: 2026/06/16 17:30:32 by vtarasov         ###   ########.fr       */
+/*   Updated: 2026/06/17 20:40:29 by vtarasov         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include "libft.h"
 
-void	*ft_memcpy(void *dest, const void *src, long unsigned int n)
+void	*ft_memcpy(void *dest, const void *src, size_t n)
 {
-	int	a;
+	unsigned int	a;
 
 	a = 0;
 	while (a < n)
